Fix: warning C4244 MSVC in Boss.cpp (conversione int->float)
PROBLEMA: MSVC /W4 segnalava warning C4244 in Boss.cpp alle righe:
  - 1401: float toothW = (size*3/5) / numTeeth;
    (size e numTeeth sono int, il risultato int viene convertito a float)
  - 1412: sf::Vector2f(size, 15.0f)
    (size e' int, Vector2f vuole float)
  + altri ~20 casi simili in renderPrimitives dove size (int) veniva
    passato a Vector2f(size*3/10, ...) o usato in espressioni float.

FIX STRUTTURALE: cambiato il tipo del membro Boss::size da 'int' a
'float' (e di conseguenza il getter getSize() da 'int' a 'float').
Questo elimina tutti i ~20 warning C4244 in un colpo solo, perche':
  - Tutte le espressioni size*3/10, size*4/5, ecc. ora sono float*int
    che produce float (non int come prima)
  - Tutte le chiamate Vector2f(size, ...) ora sono float->float
  - I valori assegnati (120, 124, 160, ...) vengono promossi a float

I punti in Game.cpp che usano boss->getSize() (in confronti float come
dx*dx + dy*dy < boss->getSize()/2 * ...) continuano a funzionare
correttamente perche' la conversione float->float e' no-op.

Rimossi i cast espliciti (float)size aggiunti precedentemente come
patch temporanea, ora non piu' necessari.

Compila senza errori ne warning (verificato con g++ -std=c++17
-fsyntax-only -Wall -Wextra -Wconversion).

+++ src/Boss.h
@@ -68,7 +68,12 @@ public:
 
     bool isDead() const { return health <= 0; }
     sf::Vector2f getPos() const { return pos; }
-    int getSize() const { return size; }
+    // FIX C4244: cambiato da int a float per evitare warning di conversione
+    // implicita int->float nelle ~20 espressioni Vector2f(size*3/10, ...) in
+    // Boss.cpp::renderPrimitives. MSVC /W4 segnala ogni conversione int->float
+    // come C4244. Avendo size come float, tutte le espressioni sono float->float.
+    // I valori assegnati (120, 124, 160, ...) vengono promossi a float.
+    float getSize() const { return size; }
     BossType getType() const { return type; }
     int getHealth() const { return health; }
     int getMaxHealth() const { return maxHealth; }
@@ -82,7 +87,8 @@ public:
 private:
     sf::Vector2f pos;
     int dx, dy;       // direzione di rimbalzo
-    int size;         // dimensione del boss (px); usata anche come raggio per le collisioni
+    // FIX C4244: cambiato da int a float (vedi getSize() sopra per dettagli)
+    float size;       // dimensione del boss (px); usata anche come raggio per le collisioni
     int health, maxHealth;
     int speed;
     BossType type;
